feat:  refacto code for better lisibility
