+++ .gitignore
@@ -15,3 +15,4 @@ bin
 lib
 data-*
 .env.local
+cache-*
