+++ src/analysis/kpi/__tests__/getVisitsOnContribWithIdcc.test.ts
@@ -1,0 +1,23 @@
+import { DataFrame } from "data-forge";
+
+import { getVisitsOnContribWithIdcc } from "../../kpi";
+
+describe("#getVisitsOnContribWithIdcc", () => {
+  it("should return df of contrib with only contrib with idcc", () => {
+    const data = [
+      { url: "https://code.travail.gouv.fr/contribution/mon-outil" },
+      { url: "https://code.travail.gouv.fr/contribution/1234" },
+      { url: "https://code.travail.gouv.fr/contribution/mon-outil1" },
+    ];
+    const dataset = new DataFrame(data);
+    const dataExpected = [
+      { url: "https://code.travail.gouv.fr/contribution/1234" },
+    ];
+    const datasetExpected = new DataFrame(dataExpected);
+    // When
+    const result = getVisitsOnContribWithIdcc(dataset);
+
+    // Then
+    expect(result).toStrictEqual(datasetExpected);
+  });
+});
